simpler projection autoHeight (#1166)
* autoHeight

the height for projections (including the null projection when Plot.geo is used) is computed for a target frame with an aspect ratio that defaults to the golden ratio. When using facets, the computation takes into account the number of rows and columns, with similar limits to what total size is acceptable (1260px).

The projection's preferred aspect ratio is adapted for a few named projections. For example "equal-earth" and "equirectangular" are wider than tall, "mercator" and a few azimuthal projections default to a square.

closes #1136
supersedes #1162

* determine the width of a facet from the (default) value of paddingInner=0.1, paddingOuter=0, instead of relying on the fx scale still being in the intermediate state where its range is [0,1]. There's a 2px discrepancy in the tests, which is probably coming from some rounding.

* refactor: use the ratio implied by tx and ty

* aspectRatio

* auto height based on width

* Update src/dimensions.js

Co-authored-by: Philippe Rivière <fil@rezo.net>

* fix auto height when geometry-less projection

* fix default height for custom projections

Co-authored-by: Mike Bostock <mbostock@gmail.com>